Create pixel-to-rem converter for certain type values
Figma only supports pixel and percentage values for line heights
and font sizes. So we're creating our own support for that here.
REM values are defined as being a multiplier of the base font
size.

Since we keep the CSS default of 16px for that base font size
in Clockwork, we can assume it here too. This *will* cause problems
if downstream platforms don't use 16px on `<body>`.